pyproject.toml
[project]
name = "minimalinferencegateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "httpx",
    "pyyaml",
]
